requirements.txt
requests
beautifulsoup4
playwright
openai
PyYAML
python-dotenv
schedule
